accessible/tests/mochitest/value/test_range.html

<html>

<head>
  <title>nsIAccessible value testing for input@type=range element</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../value.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/chrome-harness.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      // HTML5 progress element tests
      testValue("range", "50", 50, 0, 100, 1);
      testValue("range_value", "1", 1, 0, 100, 1);
      testValue("range_step", "50", 50, 0, 100, 1);
      testValue("range_min42", "71", 71, 42, 100, 1);
      testValue("range_max42", "21", 21, 0, 42, 1);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=559764"
     title="make HTML5 input@type=range element accessible">
    Bug 559764
  </a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <pre id="test">
  </pre>

  <!-- HTML5 input@type=range element -->
  <input type="range" id="range">
  <input type="range" id="range_value" value="1">
  <input type="range" id="range_step" step="1">
  <input type="range" id="range_min42" min="42">
  <input type="range" id="range_max42" max="42">
</body>
</html>
